Revert "Update index.js"
This reverts commit 5026f04da6b9fc8be3ca99f01f835e3737cb454c.

+++ src/backend/gl/texture/index.js
@@ -87,16 +87,14 @@ class GLTexture extends Texture {
   }
 
   delete() {
-    if (this._deleted) return; 
+    if (this._deleted) return;
+    this._deleted = true;
     if (this.texture._refs) {
       this.texture._refs--;
       if (this.texture._refs) return;
     }
-    this._deleted = true;
     this.context.deleteTexture(this.texture);
-    
-    // Note: It must be the case that this.texture._refs === 0
-    if (this._framebuffer) {
+    if (this.texture._refs === 0 && this._framebuffer) {
       this.context.deleteFramebuffer(this._framebuffer);
       this._framebuffer = null;
     }
